+++ CHANGES.rst
@@ -5,7 +5,8 @@
 2.1.2 (unreleased)
 ==================
 
-- Nothing changed yet.
+- Allow to delete a recurred event and put a new event with the same category
+  into its place. Before this fix the new event might not be displayed.
 
 
 2.1.1 (2017-04-09)